envoy: Call 'callback' also when using current resource version
Call the completion callback also when there is no change in the
cached resource and current version is used, but it has not been
acknowledged yet.

In current use the prior behavior did not cause a problem, but the new
Envoy CRD code depends on the callback to be called also in this case.

Note that the callback is NOT called when there is
no change in the resource and the current version has already been
acknowledged.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>