helm: Add helm option for spire server socket
This unix socket is required for cilium operator to connect to spire server
to register for new spire identity. Kindly note that accessing to spire
server must be done via unix socket only. As operator can run with HA mode,
the leader pod might not have spire server running on the same node. This
limitation will be considered and tackled later.

Signed-off-by: Tam Mach <tam.mach@cilium.io>